Updating tests after refactoring
