+++ app/controllers/application-overview.ts
@@ -3,8 +3,9 @@ import { NextFunction, Request, Response, Router } from 'express';
 import _ from 'lodash';
 import { paths } from '../paths';
 import UpdateAppealService from '../service/update-appeal-service';
-import { getAppealApplicationHistory, getAppealApplicationNextStep } from '../utils/application-state-utils';
+import { getAppealApplicationNextStep } from '../utils/application-state-utils';
 import { buildProgressBarStages } from '../utils/progress-bar-utils';
+import { getAppealApplicationHistory } from '../utils/timeline-utils';
 import { asBooleanValue } from '../utils/utils';
 
 function getAppealRefNumber(appealRef: string) {
