+++ test/unit/controllers/application-overview.test.ts
@@ -1,5 +1,4 @@
 import { NextFunction, Request, Response } from 'express';
-import moment from 'moment';
 import {
   getAppealRefNumber,
   getApplicationOverview,
@@ -124,41 +123,42 @@ describe('Confirmation Page Controller', () => {
       completed: false,
       title: 'Your appeal<br/> decision'
     } ];
-    const date = moment().format('DD MMMM YYYY');
 
-    const expectedHistory = [ {
-      'date': date,
-      'title': 'Your appeal details',
-      'text': 'You sent your appeal details to the Tribunal.',
-      'links': [
-        {
-          'href': '{{ paths.detailsViewers.appealDetails }}',
-          'text': 'Your appeal details',
-          'title': 'What you sent'
-        }, {
-          'href': '{{ paths.guidancePages.tribunalCaseworker }}',
-          'text': 'What is a Tribunal Caseworker?',
-          'title': 'Helpful information'
-        } ]
-    }, {
-      'date': date,
-      'title': 'Your appeal argument',
-      'text': 'You told us why you think the Home Office decision to refuse your claim is wrong.',
-      'links': [
-        {
-          'title': 'What you sent',
-          'text': 'Why you think the Home Office is wrong',
-          'href': '{{ paths.detailsViewers.reasonsForAppeal }}'
-        }, {
-          'title': 'Useful documents',
-          'text': 'Home Office documents about your case',
-          'href': '{{ paths.detailsViewers.homeOfficeDocuments }}'
-        }, {
-          'title': 'Helpful information',
-          'text': 'Understanding your Home Office documents',
-          'href': '{{ paths.guidancePages.homeOfficeDocuments }}'
-        } ]
-    } ];
+    const expectedHistory = {
+      appealArgumentSection: [ {
+        'date': '27 February 2020',
+        'text': 'You told us why you think the Home Office decision to refuse your claim is wrong.',
+        'links': [
+          {
+            'title': 'What you sent',
+            'text': 'Why you think the Home Office is wrong',
+            'href': '{{ paths.detailsViewers.reasonsForAppeal }}'
+          }, {
+            'title': 'Useful documents',
+            'text': 'Home Office documents about your case',
+            'href': '{{ paths.detailsViewers.homeOfficeDocuments }}'
+          }, {
+            'title': 'Helpful information',
+            'text': 'Understanding your Home Office documents',
+            'href': '{{ paths.guidancePages.homeOfficeDocuments }}'
+          } ]
+      }
+      ],
+      appealDetailsSection: [ {
+        'date': '27 February 2020',
+        'text': 'You sent your appeal details to the Tribunal.',
+        'links': [
+          {
+            'title': 'What you sent',
+            'text': 'Your appeal details',
+            'href': '{{ paths.detailsViewers.appealDetails }}'
+          }, {
+            'title': 'Helpful information',
+            'text': 'What is a Tribunal Caseworker?',
+            'href': '{{ paths.guidancePages.tribunalCaseworker }}'
+          } ]
+      } ]
+    };
 
     expect(res.render).to.have.been.calledOnce.calledWith('application-overview.njk', {
       name: 'Alex Developer',
@@ -208,7 +208,7 @@ describe('Confirmation Page Controller', () => {
       allowedAskForMoreTime: false
     };
 
-    const expectedStages = [{
+    const expectedStages = [ {
       active: true,
       ariaLabel: 'Your appeal details stage',
       completed: false,
@@ -228,13 +228,13 @@ describe('Confirmation Page Controller', () => {
       ariaLabel: 'Your appeal decision stage',
       completed: false,
       title: 'Your appeal<br/> decision'
-    }];
+    } ];
 
     expect(res.render).to.have.been.calledOnce.calledWith('application-overview.njk', {
       name: 'Alex Developer',
       appealRefNumber: undefined,
       applicationNextStep: expectedNextStep,
-      history: [],
+      history: { appealArgumentSection: [], appealDetailsSection: [] },
       stages: expectedStages,
       saved: false,
       askForMoreTimeFeatureEnabled: false
@@ -266,7 +266,7 @@ describe('Confirmation Page Controller', () => {
       info: null
     };
 
-    const expectedStages = [{
+    const expectedStages = [ {
       active: true,
       ariaLabel: 'Your appeal details stage',
       completed: false,
@@ -286,13 +286,13 @@ describe('Confirmation Page Controller', () => {
       ariaLabel: 'Your appeal decision stage',
       completed: false,
       title: 'Your appeal<br/> decision'
-    }];
+    } ];
 
     expect(res.render).to.have.been.calledOnce.calledWith('application-overview.njk', {
       name: 'Alex Developer',
       appealRefNumber: 'appealNumber',
       applicationNextStep: expectedNextStep,
-      history: [],
+      history: { appealArgumentSection: [], appealDetailsSection: [] },
       stages: expectedStages,
       saved: false,
       askForMoreTimeFeatureEnabled: false
@@ -325,7 +325,7 @@ describe('Confirmation Page Controller', () => {
       info: null
     };
 
-    const expectedStages = [{
+    const expectedStages = [ {
       active: true,
       ariaLabel: 'Your appeal details stage',
       completed: false,
@@ -345,13 +345,13 @@ describe('Confirmation Page Controller', () => {
       ariaLabel: 'Your appeal decision stage',
       completed: false,
       title: 'Your appeal<br/> decision'
-    }];
+    } ];
 
     expect(res.render).to.have.been.calledOnce.calledWith('application-overview.njk', {
       name: 'Alex Developer',
       appealRefNumber: 'RP/50004/2020',
       applicationNextStep: expectedNextStep,
-      history: [],
+      history: { appealArgumentSection: [], appealDetailsSection: [] },
       stages: expectedStages,
       saved: false,
       askForMoreTimeFeatureEnabled: false
@@ -383,7 +383,7 @@ describe('Confirmation Page Controller', () => {
     expect(result).to.equal(appealReferenceNumber);
   });
 
-  it('getAppealRefNumber with inital draft ',() => {
+  it('getAppealRefNumber with inital draft ', () => {
     const req = {
       session: {
         appeal: {
