+++ test/unit/controllers/ask-for-more-time.test.ts
@@ -7,8 +7,8 @@ import {
   postCheckAndSend,
   setupAskForMoreTimeController
 } from '../../../app/controllers/ask-for-more-time/ask-for-more-time';
+import { Events } from '../../../app/data/events';
 import { paths } from '../../../app/paths';
-import { Events } from '../../../app/service/ccd-service';
 import UpdateAppealService from '../../../app/service/update-appeal-service';
 import Logger from '../../../app/utils/logger';
 import { expect, sinon } from '../../utils/testUtils';
