+++ test/unit/controllers/contact-details.test.ts
@@ -4,8 +4,8 @@ import {
   postContactDetails,
   setupContactDetailsController
 } from '../../../app/controllers/appeal-application/contact-details';
+import { Events } from '../../../app/data/events';
 import { paths } from '../../../app/paths';
-import { Events } from '../../../app/service/ccd-service';
 import UpdateAppealService from '../../../app/service/update-appeal-service';
 import Logger from '../../../app/utils/logger';
 import i18n from '../../../locale/en.json';
