+++ test/unit/controllers/enter-address.test.ts
@@ -6,8 +6,8 @@ import {
   postManualEnterAddressPage,
   setupPersonalDetailsController
 } from '../../../app/controllers/appeal-application/personal-details';
+import { Events } from '../../../app/data/events';
 import { paths } from '../../../app/paths';
-import { Events } from '../../../app/service/ccd-service';
 import UpdateAppealService from '../../../app/service/update-appeal-service';
 import Logger from '../../../app/utils/logger';
 import { expect, sinon } from '../../utils/testUtils';
