+++ test/unit/controllers/enter-dob.test.ts
@@ -5,8 +5,8 @@ import {
   postDateOfBirth,
   setupPersonalDetailsController
 } from '../../../app/controllers/appeal-application/personal-details';
+import { Events } from '../../../app/data/events';
 import { paths } from '../../../app/paths';
-import { Events } from '../../../app/service/ccd-service';
 import UpdateAppealService from '../../../app/service/update-appeal-service';
 import Logger from '../../../app/utils/logger';
 import i18n from '../../../locale/en.json';
