+++ test/unit/controllers/nationatily-page.test.ts
@@ -7,8 +7,8 @@ import {
   setupPersonalDetailsController
 } from '../../../app/controllers/appeal-application/personal-details';
 import { countryList } from '../../../app/data/country-list';
+import { Events } from '../../../app/data/events';
 import { paths } from '../../../app/paths';
-import { Events } from '../../../app/service/ccd-service';
 import UpdateAppealService from '../../../app/service/update-appeal-service';
 import Logger from '../../../app/utils/logger';
 import { getNationalitiesOptions } from '../../../app/utils/nationalities';
