+++ test/unit/controllers/reasons-for-appeal/check-and-send.test.ts
@@ -4,8 +4,8 @@ import {
   postCheckAndSend,
   setupCheckAndSendController
 } from '../../../../app/controllers/reasons-for-appeal/check-and-send';
+import { Events } from '../../../../app/data/events';
 import { paths } from '../../../../app/paths';
-import { Events } from '../../../../app/service/ccd-service';
 import UpdateAppealService from '../../../../app/service/update-appeal-service';
 import { addSummaryRow, Delimiter } from '../../../../app/utils/summary-list';
 import i18n from '../../../../locale/en.json';
