+++ test/unit/controllers/type-of-appeal.test.ts
@@ -5,8 +5,8 @@ import {
   setupTypeOfAppealController
 } from '../../../app/controllers/appeal-application/type-of-appeal';
 import { appealTypes } from '../../../app/data/appeal-types';
+import { Events } from '../../../app/data/events';
 import { paths } from '../../../app/paths';
-import { Events } from '../../../app/service/ccd-service';
 import UpdateAppealService from '../../../app/service/update-appeal-service';
 import Logger from '../../../app/utils/logger';
 import { expect, sinon } from '../../utils/testUtils';
